heron step

+++ src/main/scala/xieyuheng/bebop/Propagator.scala
@@ -3,7 +3,7 @@ package xieyuheng.bebop
 import akka.actor.ActorRef
 
 case class Propagator(actor: ActorRef) {
-  def updateArg[E](c: E, n: Int): Unit = {
+  def updateArg[E](c: E, n: Double): Unit = {
     actor ! (c, n)
   }
 }
